Docs: Upstream fixes made to Neutrino's webpack-chain docs
Neutrino now has a slightly modified copy of the webpack-chain docs
in its own repository:
https://github.com/mozilla-neutrino/neutrino-dev/blob/master/docs/webpack-chain.md

As part of that import, a few docs cleanups occurred that are also
applicable to the docs in the webpack-chain repo. Upstreaming these
reduces the size of the diff so should also make it easier to keep
the two in sync in the future.