Validate that Plugin 'args' is an array
Previously if `args` was set to something other than an array via any
of `.use()`, `.tap()`, `.merge()` or `.set()`, that call would succeed
but the later `.toConfig()` would fail with an unhelpful error message
(when the `init` function tried to spread `args`).

Now validation is performed at the time `args` is set, with a more
helpful error message and stack trace.

Fixes #121.