Drop documented support for webpack 3 and earlier
We didn't test on anything other than webpack 4, so it's not clear what
functionality still worked anyway. We'll also likely be needing to make
a number of changes to webpack-chain to support webpack 5, so the fewer
old versions of webpack we have to think about, the better.

webpack-chain doesn't have a peer dependency on webpack or we'd also
adjust that (and for some use-cases a peer dep perhaps isn't appropriate,
so I think we should leave that as is for now).